Sad progress on growing bugs
Managed to remove the recursion bug. However the shrinking of the tail persists. This build is probably the closest to having it fixed, but I need to take a look at the pointIndex logic with a clear mind. The bug in this build currently is all tailBalls collapsing to the same position as the head